move text_part_obj.dart, text_obj.dart out from src folder
