+++ lib/tagtly.dart
@@ -1,4 +1,4 @@
 export 'package:tagtly/widgets/tagtly_text_form.dart';
 export 'package:tagtly/widgets/tagtly_text_view.dart';
-export 'package:tagtly/src/text_obj.dart';
+export 'package:tagtly/text_obj.dart';
 export 'package:tagtly/src/text_enum.dart';
