+++ lib/text_part_obj.dart
